Fix DB reading bug.
Doubles were getting cast to longs accidentally. Fixed and updated test to catch it.

+++ OpenScienceJournal/whistlepunk_library/src/main/java/com/google/android/apps/forscience/whistlepunk/sensordb/SensorDatabaseImpl.java
@@ -228,7 +228,7 @@ public class SensorDatabaseImpl implements SensorDatabase {
                             while (c.moveToNext()) {
                                 long timeStamp = c.getLong(0);
                                 observableEmitter.onNext(new ScalarReading(timeStamp,
-                                        c.getLong(1), c.getString(2)));
+                                        c.getDouble(1), c.getString(2)));
                                 mLastTimeStampWritten = timeStamp;
                                 count++;
                             }
